+++ shaders/lib/uniforms.glsl
@@ -126,6 +126,7 @@ uniform vec3 previousCameraPositionFract;
 
     uniform mat4 dhProjection;
     uniform mat4 dhProjectionInverse;
+    uniform mat4 dhPreviousProjection;
     
     uniform sampler2D dhDepthTex;
     uniform sampler2D dhDepthTex1;